Add line numbers to show Code Lens
By providing the line number on which the test/feature was defined we allow Test Explorer UI to show Code Len to be able to run the test right from where it was defined.

This also allow the Show Source button to go directly to the test line inside the file.